Remove NAPI_EXPERIMENTAL as it was causing
```
FATAL ERROR: Finalizer is calling a function that may affect GC state.
The finalizers are run directly from GC and must not affect GC state.
Use `node_api_post_finalizer` from inside of the finalizer to work around this issue.
It schedules the call as a new task in the event loop.
```
with some node versions